trading_app/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Bootstrap==3.3.7.1
Jinja2==3.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
pandas==1.5.3
numpy==1.23.3
pyotp==2.8.1
pytz==2023.7
requests==2.26.0
python-dotenv==0.19.0
pyyaml==6.0
toml==0.12.2
yattag==1.17.2
smartapi-python==0.2.5